requirements.txt
myo-python==1.0.3
numpy==1.14.5
